C3
- added localStorage for the watched movie list
- added automatic focus for the search input with useRef
- added autofocus on enter key too
- added a new background variable to check user rating decisions with useRef
- created a new custom hook (useMovies) for the movie fetch
- created another custom hook (useLocalStorage) to handle local storage getter and setter globally
- created a custom hook (useKey) to handle key press actions globally (like hit enter or escape)
