+++ src/components/common/useKey.js
@@ -1,0 +1,20 @@
+import { useEffect } from "react";
+
+export function useKey(key, action) {
+  useEffect(
+    function () {
+      function callback(e) {
+        if (e.code.toLowerCase() === key.toLowerCase()) {
+          action();
+        }
+      }
+      document.addEventListener("keydown", callback);
+
+      //cleanup: remove event listener
+      return function () {
+        document.removeEventListener("keydown", callback);
+      };
+    },
+    [action, key]
+  );
+}
