+++ src/components/navbar/Search.js
@@ -1,6 +1,18 @@
-import { useState } from "react";
+import { useRef } from "react";
+import { useKey } from "../common/useKey";
 
 export default function Search({ query, setQuery }) {
+  const inputEl = useRef(null);
+
+  // using my custom hook here to handle key action globally, added a longer action param to it
+  useKey("Enter", function () {
+    if (document.activeElement === inputEl.current) {
+      return;
+    }
+    inputEl.current.focus();
+    setQuery("");
+  });
+
   return (
     <input
       className="search"
@@ -8,6 +20,7 @@ export default function Search({ query, setQuery }) {
       placeholder="Search movies..."
       value={query}
       onChange={(e) => setQuery(e.target.value)}
+      ref={inputEl}
     />
   );
 }
